moved wiki_cache helper to testing module and refactored to allow passing :memory:

+++ lib/wiki_nodes/http/cache.py
@@ -5,12 +5,8 @@ Caching for wiki pages, images, searches, etc.
 from __future__ import annotations
 
 import logging
-import pickle
-from datetime import datetime
 from pathlib import Path
 from typing import TYPE_CHECKING, Optional, Union, Any, Mapping
-
-from requests import Response
 
 from db_cache import TTLDBCache, DBCache
 from db_cache.utils import get_user_cache_dir
@@ -36,11 +32,9 @@ class WikiCache:
 
     def __init__(self, host: str, ttl: int = 21_600, base_dir: PathLike = None, img_dir: PathLike = None):
         self.ttl = ttl  # Note: default value of 21_600 = 3600 * 6 (6 hours)
-        self.base_dir = Path(base_dir or get_user_cache_dir(f'wiki/{host}'))
-        self.base_dir.mkdir(parents=True, exist_ok=True)
+        self.base_dir = _prep_dir(base_dir, f'wiki/{host}')
         self.reset_caches(False)
-        self.img_dir = Path(img_dir or get_user_cache_dir(f'wiki/{host}/images'))
-        self.img_dir.mkdir(parents=True, exist_ok=True)
+        self.img_dir = _prep_dir(img_dir, f'wiki/{host}/images')
 
     def __getstate__(self) -> dict[str, Union[int, Path]]:
         return {'ttl': self.ttl, 'base_dir': self.base_dir, 'img_dir': self.img_dir}
@@ -53,18 +47,20 @@ class WikiCache:
 
     def reset_caches(self, hard: bool = False):
         cache_dir = self.base_dir
-        if hard:
-            for path in cache_dir.iterdir():
-                if path.is_file() and path.suffix == '.db':
+        if hard and cache_dir != ':memory:':
+            for path in cache_dir.glob('*.db'):
+                if path.is_file():
                     log.debug(f'Deleting cache file: {path.as_posix()}')
                     path.unlink()
 
-        self.pages = TTLDBCache('pages', cache_dir=cache_dir, ttl=self.ttl)
-        self.search_titles = TTLDBCache('search_titles', cache_dir=cache_dir, ttl=self.ttl)
-        self.searches = TTLDBCache('searches', cache_dir=cache_dir, ttl=self.ttl)
+        key = 'db_path' if cache_dir == ':memory:' else 'cache_dir'
+        kwargs = {key: cache_dir, 'ttl': self.ttl}
+        self.pages = TTLDBCache('pages', **kwargs)
+        self.search_titles = TTLDBCache('search_titles', **kwargs)
+        self.searches = TTLDBCache('searches', **kwargs)
         # All keys in normalized_titles should be normalized to upper case to improve matching and prevent dupes
-        self.normalized_titles = DBCache('normalized_titles', cache_dir=cache_dir, time_fmt='%Y')
-        self.misc = TTLDBCache('misc', cache_dir=cache_dir, ttl=self.ttl)
+        self.normalized_titles = DBCache('normalized_titles', time_fmt='%Y', **{key: cache_dir})
+        self.misc = TTLDBCache('misc', **kwargs)
 
     def get_misc(self, group: str, titles: Titles) -> tuple[list[str], dict[str, Any]]:
         titles = [titles] if isinstance(titles, str) else titles
@@ -93,3 +89,13 @@ class WikiCache:
     def store_image(self, name: Optional[str], data: bytes):
         if name:
             self.img_dir.joinpath(name).write_bytes(data)
+
+
+def _prep_dir(path: PathLike, default: str) -> Union[Path, str]:
+    if path:
+        if path == ':memory:':
+            return path
+        path = Path(path)
+        path.mkdir(parents=True, exist_ok=True)
+        return path
+    return get_user_cache_dir(default)
